+++ tests/roots/test-default/conf.py
@@ -1,0 +1,9 @@
+"""This configuration is cases for efault behavior."""
+extensions = [
+    "sphinx_revealjs",
+]
+
+# To skip toctree
+rst_prolog = """
+:orphan:
+"""