+++ tests/roots/test-for-config/conf.py
@@ -1,0 +1,10 @@
+"""This configuration is cases for changing conf.py
+"""
+extensions = [
+    "sphinx_revealjs",
+]
+
+# To skip toctree
+rst_prolog = """
+:orphan:
+"""