+++ sphinx_revealjs/__init__.py
@@ -1,6 +1,6 @@
 """Root module for sphinx-revealjs."""
 
-__version__ = "2.0.0"
+__version__ = "2.0.1"
 
 import sys